requirements.txt
fastapi~=0.115.12
aiogram~=3.20.0.post0
uvicorn
gspread~=6.2.0
oauth2client~=4.1.3
pytz
python-dotenv~=1.1.0